+++ src/main/java/org/gridsuite/study/server/controller/StudyController.java
@@ -6,10 +6,8 @@
  */
 package org.gridsuite.study.server.controller;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.powsybl.iidm.network.ThreeSides;
-import com.powsybl.timeseries.DoubleTimeSeries;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
@@ -20,14 +18,11 @@ import org.gridsuite.filter.globalfilter.GlobalFilter;
 import org.gridsuite.filter.utils.EquipmentType;
 import org.gridsuite.study.server.StudyApi;
 import org.gridsuite.study.server.dto.*;
-import org.gridsuite.study.server.dto.dynamicsimulation.event.EventInfos;
 import org.gridsuite.study.server.dto.elasticsearch.EquipmentInfos;
 import org.gridsuite.study.server.dto.modification.*;
 import org.gridsuite.study.server.dto.networkexport.ExportNetworkStatus;
 import org.gridsuite.study.server.dto.networkexport.NodeExportInfos;
 import org.gridsuite.study.server.dto.sequence.NodeSequenceType;
-import org.gridsuite.study.server.dto.timeseries.TimeSeriesMetadataInfos;
-import org.gridsuite.study.server.dto.timeseries.TimelineEventInfos;
 import org.gridsuite.study.server.elasticsearch.EquipmentInfosService;
 import org.gridsuite.study.server.error.StudyException;
 import org.gridsuite.study.server.exception.PartialResultException;
@@ -49,7 +44,6 @@ import java.beans.PropertyEditorSupport;
 import java.util.*;
 
 import static org.gridsuite.study.server.StudyConstants.*;
-import static org.gridsuite.study.server.dto.ComputationType.*;
 import static org.gridsuite.study.server.error.StudyBusinessErrorCode.MOVE_NETWORK_MODIFICATION_FORBIDDEN;
 
 /**
@@ -68,7 +62,6 @@ public class StudyController {
     private final CaseService caseService;
     private final RemoteServicesInspector remoteServicesInspector;
     private final RootNetworkService rootNetworkService;
-    private final RootNetworkNodeInfoService rootNetworkNodeInfoService;
     private final RebuildNodeService rebuildNodeService;
 
     public StudyController(StudyService studyService,
@@ -79,7 +72,6 @@ public class StudyController {
                            CaseService caseService,
                            RemoteServicesInspector remoteServicesInspector,
                            RootNetworkService rootNetworkService,
-                           RootNetworkNodeInfoService rootNetworkNodeInfoService,
                            RebuildNodeService rebuildNodeService) {
         this.studyService = studyService;
         this.networkModificationTreeService = networkModificationTreeService;
@@ -89,7 +81,6 @@ public class StudyController {
         this.caseService = caseService;
         this.remoteServicesInspector = remoteServicesInspector;
         this.rootNetworkService = rootNetworkService;
-        this.rootNetworkNodeInfoService = rootNetworkNodeInfoService;
         this.rebuildNodeService = rebuildNodeService;
     }
 
@@ -821,27 +812,6 @@ public class StudyController {
         return ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON).body(studyService.getResultEnumValues(nodeUuid, rootNetworkUuid, computingType, enumName));
     }
 
-    @GetMapping(value = "/studies/{studyUuid}/dynamic-simulation/provider")
-    @Operation(summary = "Get dynamic simulation provider for a specified study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The dynamic simulation provider is returned")})
-    public ResponseEntity<String> getDynamicSimulationProvider(@PathVariable("studyUuid") UUID studyUuid) {
-        return ResponseEntity.ok().body(studyService.getDynamicSimulationProvider(studyUuid));
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/dynamic-security-analysis/provider")
-    @Operation(summary = "Get dynamic security analysis provider for a specified study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The dynamic security analysis provider is returned")})
-    public ResponseEntity<String> getDynamicSecurityAnalysisProvider(@PathVariable("studyUuid") UUID studyUuid) {
-        return ResponseEntity.ok().body(studyService.getDynamicSecurityAnalysisProvider(studyUuid));
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/dynamic-margin-calculation/provider")
-    @Operation(summary = "Get dynamic margin calculation provider for a specified study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The dynamic margin calculation provider is returned")})
-    public ResponseEntity<String> getDynamicMarginCalculationProvider(@PathVariable("studyUuid") UUID studyUuid) {
-        return ResponseEntity.ok().body(studyService.getDynamicMarginCalculationProvider(studyUuid));
-    }
-
     @PostMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/network/substations/{substationId}/svg")
     @Operation(summary = "get the substation diagram for the given network and substation")
     @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The svg"),
@@ -1371,267 +1341,6 @@ public class StudyController {
         studyService.notify(studyUuid);
         return ResponseEntity.ok().build();
     }
-
-    // --- Dynamic Simulation Endpoints BEGIN --- //
-
-    @PostMapping(value = "/studies/{studyUuid}/dynamic-simulation/parameters")
-    @Operation(summary = "Set dynamic simulation parameters on study, reset to default ones if empty body")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The dynamic simulation parameters are set")})
-    public ResponseEntity<Void> setDynamicSimulationParameters(
-            @PathVariable("studyUuid") UUID studyUuid,
-            @RequestBody(required = false) String dsParameter,
-            @RequestHeader(HEADER_USER_ID) String userId) {
-        studyService.setDynamicSimulationParameters(studyUuid, dsParameter, userId);
-        return ResponseEntity.ok().build();
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/dynamic-simulation/parameters")
-    @Operation(summary = "Get dynamic simulation parameters on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The dynamic simulation parameters")})
-    public ResponseEntity<String> getDynamicSimulationParameters(
-            @PathVariable("studyUuid") UUID studyUuid) {
-        return ResponseEntity.ok().body(studyService.getDynamicSimulationParameters(studyUuid));
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/nodes/{nodeUuid}/dynamic-simulation/events")
-    @Operation(summary = "Get dynamic simulation events from a node")
-    @ApiResponses(value = {
-        @ApiResponse(responseCode = "200", description = "The dynamic simulation events was returned"),
-        @ApiResponse(responseCode = "404", description = "The study/node is not found")})
-    public ResponseEntity<List<EventInfos>> getDynamicSimulationEvents(@Parameter(description = "Study UUID") @PathVariable("studyUuid") UUID studyUuid,
-                                                                       @Parameter(description = "Node UUID") @PathVariable("nodeUuid") UUID nodeUuid) {
-        List<EventInfos> dynamicSimulationEvents = studyService.getDynamicSimulationEvents(nodeUuid);
-        return ResponseEntity.ok().body(dynamicSimulationEvents);
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/nodes/{nodeUuid}/dynamic-simulation/events", params = {"equipmentId"})
-    @Operation(summary = "Get dynamic simulation event from a node with a given equipment id")
-    @ApiResponses(value = {
-        @ApiResponse(responseCode = "200", description = "The dynamic simulation event was returned"),
-        @ApiResponse(responseCode = "404", description = "The study/node is not found")})
-    public ResponseEntity<EventInfos> getDynamicSimulationEvent(@Parameter(description = "Study UUID") @PathVariable("studyUuid") UUID studyUuid,
-                                                               @Parameter(description = "Node UUID") @PathVariable("nodeUuid") UUID nodeUuid,
-                                                               @Parameter(description = "Equipment id") @RequestParam(value = "equipmentId") String equipmentId) {
-        EventInfos dynamicSimulationEvent = studyService.getDynamicSimulationEvent(nodeUuid, equipmentId);
-        return dynamicSimulationEvent != null ? ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON).body(dynamicSimulationEvent) :
-                ResponseEntity.noContent().build();
-    }
-
-    @PostMapping(value = "/studies/{studyUuid}/nodes/{nodeUuid}/dynamic-simulation/events")
-    @Operation(summary = "Create a dynamic simulation event for a node")
-    @ApiResponses(value = {
-        @ApiResponse(responseCode = "200", description = "The network event was created"),
-        @ApiResponse(responseCode = "404", description = "The study/node is not found")})
-    public ResponseEntity<Void> createDynamicSimulationEvent(@Parameter(description = "Study UUID") @PathVariable("studyUuid") UUID studyUuid,
-                                                             @Parameter(description = "Node UUID") @PathVariable("nodeUuid") UUID nodeUuid,
-                                                             @RequestBody EventInfos event,
-                                                             @RequestHeader(HEADER_USER_ID) String userId) {
-        studyService.assertCanUpdateNodeInStudy(studyUuid, nodeUuid);
-        studyService.createDynamicSimulationEvent(studyUuid, nodeUuid, userId, event);
-        return ResponseEntity.ok().build();
-    }
-
-    @PutMapping(value = "/studies/{studyUuid}/nodes/{nodeUuid}/dynamic-simulation/events")
-    @Operation(summary = "Update a dynamic simulation event for a node")
-    @ApiResponses(value = {
-        @ApiResponse(responseCode = "200", description = "The dynamic simulation event was updated"),
-        @ApiResponse(responseCode = "404", description = "The study/node is not found")})
-    public ResponseEntity<Void> updateDynamicSimulationEvent(@Parameter(description = "Study UUID") @PathVariable("studyUuid") UUID studyUuid,
-                                                             @Parameter(description = "Node UUID") @PathVariable("nodeUuid") UUID nodeUuid,
-                                                             @RequestBody EventInfos event,
-                                                             @RequestHeader(HEADER_USER_ID) String userId) {
-        studyService.assertCanUpdateNodeInStudy(studyUuid, nodeUuid);
-        studyService.updateDynamicSimulationEvent(studyUuid, nodeUuid, userId, event);
-        return ResponseEntity.ok().build();
-    }
-
-    @DeleteMapping(value = "/studies/{studyUuid}/nodes/{nodeUuid}/dynamic-simulation/events")
-    @Operation(summary = "Delete dynamic simulation events for a node")
-    @ApiResponses(value = {
-        @ApiResponse(responseCode = "200", description = "The dynamic simulation events was deleted"),
-        @ApiResponse(responseCode = "404", description = "The study/node is not found")})
-    public ResponseEntity<Void> deleteDynamicSimulationEvents(@Parameter(description = "Study UUID") @PathVariable("studyUuid") UUID studyUuid,
-                                                              @Parameter(description = "Node UUID") @PathVariable("nodeUuid") UUID nodeUuid,
-                                                              @Parameter(description = "Dynamic simulation event UUIDs") @RequestParam("eventUuids") List<UUID> eventUuids,
-                                                              @RequestHeader(HEADER_USER_ID) String userId) {
-        studyService.assertCanUpdateNodeInStudy(studyUuid, nodeUuid);
-        studyService.deleteDynamicSimulationEvents(studyUuid, nodeUuid, userId, eventUuids);
-        return ResponseEntity.ok().build();
-    }
-
-    @PostMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/dynamic-simulation/run")
-    @Operation(summary = "run dynamic simulation on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The dynamic simulation has started")})
-    public ResponseEntity<Void> runDynamicSimulation(@Parameter(description = "studyUuid") @PathVariable("studyUuid") UUID studyUuid,
-                                                     @Parameter(description = "rootNetworkUuid") @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                     @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid,
-                                                     @Parameter(description = "debug") @RequestParam(name = "debug", required = false, defaultValue = "false") boolean debug,
-                                                     @RequestHeader(HEADER_USER_ID) String userId) {
-        studyService.assertIsNodeNotReadOnly(nodeUuid);
-        studyService.assertOnQuotasAvailability(DYNAMIC_SIMULATION, userId);
-        studyService.assertCanRunOnConstructionNode(studyUuid, nodeUuid, List.of(DYNAWO_PROVIDER), studyService::getDynamicSimulationProvider);
-        studyService.runDynamicSimulation(studyUuid, nodeUuid, rootNetworkUuid, userId, debug);
-        return ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON).build();
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/dynamic-simulation/result/timeseries/metadata")
-    @Operation(summary = "Get list of time series metadata of dynamic simulation result on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "Time series metadata of dynamic simulation result"),
-        @ApiResponse(responseCode = "204", description = "No dynamic simulation metadata"),
-        @ApiResponse(responseCode = "404", description = "The dynamic simulation has not been found")})
-    public ResponseEntity<List<TimeSeriesMetadataInfos>> getDynamicSimulationTimeSeriesMetadata(@Parameter(description = "study UUID") @PathVariable("studyUuid") UUID studyUuid,
-                                                                                                @Parameter(description = "rootNetworkUuid") @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                                                                @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid) {
-        List<TimeSeriesMetadataInfos> result = rootNetworkNodeInfoService.getDynamicSimulationTimeSeriesMetadata(nodeUuid, rootNetworkUuid);
-        return CollectionUtils.isEmpty(result) ? ResponseEntity.noContent().build() :
-                ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON).body(result);
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/dynamic-simulation/result/timeseries")
-    @Operation(summary = "Get all time series of dynamic simulation result on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "All time series of dynamic simulation result"),
-        @ApiResponse(responseCode = "204", description = "No dynamic simulation timeseries"),
-        @ApiResponse(responseCode = "404", description = "The dynamic simulation has not been found")})
-    public ResponseEntity<List<DoubleTimeSeries>> getDynamicSimulationTimeSeriesResult(@Parameter(description = "study UUID") @PathVariable("studyUuid") UUID studyUuid,
-                                                                                       @Parameter(description = "rootNetworkUuid") @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                                                       @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid,
-                                                                                       @Parameter(description = "timeSeriesNames") @RequestParam(name = "timeSeriesNames",
-                                                                                               required = false) List<String> timeSeriesNames) {
-        List<DoubleTimeSeries> result = rootNetworkNodeInfoService.getDynamicSimulationTimeSeries(nodeUuid, rootNetworkUuid, timeSeriesNames);
-        return CollectionUtils.isEmpty(result) ? ResponseEntity.noContent().build() :
-                ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON).body(result);
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/dynamic-simulation/result/timeline")
-    @Operation(summary = "Get timeline events of dynamic simulation result on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "Timeline events of dynamic simulation result"),
-        @ApiResponse(responseCode = "204", description = "No dynamic simulation timeline events"),
-        @ApiResponse(responseCode = "404", description = "The dynamic simulation has not been found")})
-    public ResponseEntity<List<TimelineEventInfos>> getDynamicSimulationTimelineResult(@Parameter(description = "study UUID") @PathVariable("studyUuid") UUID studyUuid,
-                                                                                       @Parameter(description = "rootNetworkUuid") @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                                                       @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid) {
-        List<TimelineEventInfos> result = rootNetworkNodeInfoService.getDynamicSimulationTimeline(nodeUuid, rootNetworkUuid);
-        return CollectionUtils.isEmpty(result) ? ResponseEntity.noContent().build() :
-                ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON).body(result);
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/dynamic-simulation/status")
-    @Operation(summary = "Get the status of dynamic simulation result on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The status of dynamic simulation result"),
-        @ApiResponse(responseCode = "204", description = "No dynamic simulation status"),
-        @ApiResponse(responseCode = "404", description = "The dynamic simulation has not been found")})
-    public ResponseEntity<String> getDynamicSimulationStatus(@Parameter(description = "study UUID") @PathVariable("studyUuid") UUID studyUuid,
-                                                             @Parameter(description = "rootNetworkUuid") @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                             @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid) {
-        String result = rootNetworkNodeInfoService.getDynamicSimulationStatus(nodeUuid, rootNetworkUuid);
-        return result != null ? ResponseEntity.ok().body(result) : ResponseEntity.noContent().build();
-    }
-
-    // --- Dynamic Simulation Endpoints END --- //
-
-    // --- Dynamic Security Analysis Endpoints BEGIN --- //
-
-    @PostMapping(value = "/studies/{studyUuid}/dynamic-security-analysis/parameters")
-    @Operation(summary = "Set dynamic security analysis parameters on study, reset to default one if empty body")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The dynamic security analysis parameters are set")})
-    public ResponseEntity<Void> setDynamicSecurityAnalysisParameters(
-            @PathVariable("studyUuid") UUID studyUuid,
-            @RequestBody(required = false) String dsaParameter,
-            @RequestHeader(HEADER_USER_ID) String userId) {
-        return studyService.setDynamicSecurityAnalysisParameters(studyUuid, dsaParameter, userId) ?
-                ResponseEntity.noContent().build() :
-                ResponseEntity.ok().build();
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/dynamic-security-analysis/parameters")
-    @Operation(summary = "Get dynamic security analysis parameters on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The dynamic security analysis parameters")})
-    public ResponseEntity<String> getDynamicSecurityAnalysisParameters(
-            @PathVariable("studyUuid") UUID studyUuid) {
-        return ResponseEntity.ok().body(studyService.getDynamicSecurityAnalysisParameters(studyUuid));
-    }
-
-    @PostMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/dynamic-security-analysis/run")
-    @Operation(summary = "run dynamic security analysis on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The dynamic security analysis has started")})
-    public ResponseEntity<Void> runDynamicSecurityAnalysis(@Parameter(description = "studyUuid") @PathVariable("studyUuid") UUID studyUuid,
-                                                     @Parameter(description = "root network id") @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                     @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid,
-                                                     @Parameter(description = "debug") @RequestParam(name = "debug", required = false, defaultValue = "false") boolean debug,
-                                                     @RequestHeader(HEADER_USER_ID) String userId) {
-        studyService.assertIsNodeNotReadOnly(nodeUuid);
-        studyService.assertOnQuotasAvailability(DYNAMIC_SECURITY_ANALYSIS, userId);
-        studyService.assertCanRunOnConstructionNode(studyUuid, nodeUuid, List.of(DYNAWO_PROVIDER), studyService::getDynamicSecurityAnalysisProvider);
-        studyService.runDynamicSecurityAnalysis(studyUuid, nodeUuid, rootNetworkUuid, userId, debug);
-        return ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON).build();
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/dynamic-security-analysis/status")
-    @Operation(summary = "Get the status of dynamic security analysis result on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The status of dynamic security analysis result"),
-        @ApiResponse(responseCode = "204", description = "No dynamic security analysis status"),
-        @ApiResponse(responseCode = "404", description = "The dynamic security analysis has not been found")})
-    public ResponseEntity<String> getDynamicSecurityAnalysisStatus(@Parameter(description = "study UUID") @PathVariable("studyUuid") UUID studyUuid,
-                                                                                          @Parameter(description = "root network id") @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                                                          @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid) {
-        String result = rootNetworkNodeInfoService.getDynamicSecurityAnalysisStatus(nodeUuid, rootNetworkUuid);
-        return result != null ? ResponseEntity.ok().body(result) : ResponseEntity.noContent().build();
-    }
-
-    // --- Dynamic Security Analysis Endpoints END --- //
-
-    // --- Dynamic Margin Calculation Endpoints BEGIN --- //
-
-    @PostMapping(value = "/studies/{studyUuid}/dynamic-margin-calculation/parameters")
-    @Operation(summary = "Set dynamic margin calculation parameters on study, reset to default one if empty body")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The dynamic margin calculation parameters are set")})
-    public ResponseEntity<Void> setDynamicMarginCalculationParameters(
-            @PathVariable("studyUuid") UUID studyUuid,
-            @RequestBody(required = false) String dmcParameter,
-            @RequestHeader(HEADER_USER_ID) String userId) {
-        return studyService.setDynamicMarginCalculationParameters(studyUuid, dmcParameter, userId) ?
-                ResponseEntity.noContent().build() :
-                ResponseEntity.ok().build();
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/dynamic-margin-calculation/parameters")
-    @Operation(summary = "Get dynamic margin calculation parameters on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The dynamic margin calculation parameters")})
-    public ResponseEntity<String> getDynamicMarginCalculationParameters(
-            @PathVariable("studyUuid") UUID studyUuid,
-            @RequestHeader(HEADER_USER_ID) String userId
-    ) {
-        return ResponseEntity.ok().body(studyService.getDynamicMarginCalculationParameters(studyUuid, userId));
-    }
-
-    @PostMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/dynamic-margin-calculation/run")
-    @Operation(summary = "run dynamic margin calculation on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The dynamic margin calculation has started")})
-    public ResponseEntity<Void> runDynamicMarginCalculation(@Parameter(description = "studyUuid") @PathVariable("studyUuid") UUID studyUuid,
-                                                     @Parameter(description = "root network id") @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                     @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid,
-                                                     @Parameter(description = "debug") @RequestParam(name = "debug", required = false, defaultValue = "false") boolean debug,
-                                                     @RequestHeader(HEADER_USER_ID) String userId) throws JsonProcessingException {
-        studyService.assertIsNodeNotReadOnly(nodeUuid);
-        studyService.assertOnQuotasAvailability(DYNAMIC_MARGIN_CALCULATION, userId);
-        studyService.assertCanRunOnConstructionNode(studyUuid, nodeUuid, List.of(DYNAWO_PROVIDER), studyService::getDynamicMarginCalculationProvider);
-        studyService.runDynamicMarginCalculation(studyUuid, nodeUuid, rootNetworkUuid, userId, debug);
-        return ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON).build();
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/dynamic-margin-calculation/status")
-    @Operation(summary = "Get the status of dynamic margin calculation result on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The status of dynamic margin calculation result"),
-        @ApiResponse(responseCode = "204", description = "No dynamic margin calculation status"),
-        @ApiResponse(responseCode = "404", description = "The dynamic margin calculation has not been found")})
-    public ResponseEntity<String> getDynamicMarginCalculationStatus(@Parameter(description = "study UUID") @PathVariable("studyUuid") UUID studyUuid,
-                                                                                          @Parameter(description = "root network id") @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                                                          @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid) {
-        String result = rootNetworkNodeInfoService.getDynamicMarginCalculationStatus(nodeUuid, rootNetworkUuid);
-        return result != null ? ResponseEntity.ok().body(result) : ResponseEntity.noContent().build();
-    }
-
-    // --- Dynamic Margin Calculation Endpoints END --- //
 
     @GetMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/network-modifications/voltage-init", produces = MediaType.APPLICATION_JSON_VALUE)
     @Operation(summary = "Get the voltage init modifications from a node")
